Moved internal.TestDB to internal/qtest
