+++ .vscode/settings.json
@@ -106,6 +106,7 @@
         "protobuf",
         "protoc",
         "protoimpl",
+        "qtest",
         "quickfeed",
         "repotoken",
         "run",
