+++ database/gormdb_assignment_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	pb "github.com/autograde/quickfeed/ag"
-	"github.com/autograde/quickfeed/internal"
+	"github.com/autograde/quickfeed/internal/qtest"
 	"github.com/autograde/quickfeed/kit/score"
 	"github.com/google/go-cmp/cmp"
 	"google.golang.org/protobuf/proto"
@@ -14,7 +14,7 @@ import (
 )
 
 func TestGormDBGetAssignment(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	if _, err := db.GetAssignmentsByCourse(10, false); err != gorm.ErrRecordNotFound {
@@ -27,7 +27,7 @@ func TestGormDBGetAssignment(t *testing.T) {
 }
 
 func TestGormDBCreateAssignmentNoRecord(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	assignment := pb.Assignment{
@@ -42,7 +42,7 @@ func TestGormDBCreateAssignmentNoRecord(t *testing.T) {
 }
 
 func TestGormDBCreateAssignment(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	user := createFakeUser(t, db, 10)
@@ -78,7 +78,7 @@ func TestGormDBCreateAssignment(t *testing.T) {
 }
 
 func TestUpdateAssignment(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	course := &pb.Course{}
@@ -142,7 +142,7 @@ func TestUpdateAssignment(t *testing.T) {
 }
 
 func TestGetAssignmentsWithSubmissions(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	// create teacher, course, user (student) and assignment
@@ -232,7 +232,7 @@ func TestGetAssignmentsWithSubmissions(t *testing.T) {
 }
 
 func TestUpdateBenchmarks(t *testing.T) {
-	db, cleanup := internal.TestDB(t)
+	db, cleanup := qtest.TestDB(t)
 	defer cleanup()
 
 	course := &pb.Course{}
